+++ res/html/recordJudging.js
@@ -1,7 +1,7 @@
-window.onload = () => {
+window.addEventListener('DOMContentLoaded', () => {
     const detail = document.getElementById('detail');
     window.onmessage = event => {
         const message = event.data;
         detail.innerHTML = message.status_html;
     };
-};
+});