+++ src/pWeb.ts
@@ -8,6 +8,7 @@ import settings from './settings';
 export default class {
     private _panel: vscode.WebviewPanel;
     private _extensionPath: string;
+    private _tempFiles: string[] = [];
 
     constructor(extensionPath: string, pid: number, tid?: string) {
         outputChannel.trace('[pWeb    ]', '"constructor"', arguments);
@@ -36,6 +37,7 @@ export default class {
                     vscode.commands.executeCommand('cyezoi.openP', pid);
             }
         });
+        this._panel.onDidDispose(this.cleanup);
 
         this.fetchData(pid, tid);
     }
@@ -63,6 +65,8 @@ export default class {
             { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'indent-fold.min.js'] },
             { 'path': ['res', 'libs', 'codemirror', 'addon', 'section', 'active-line.min.js'] },
             { 'path': ['res', 'libs', 'codemirror', 'addon', 'display', 'autorefresh.min.js'] },
+            { 'path': ['res', 'libs', 'pdfjs', 'pdf_viewer.min.css'] },
+            { 'path': ['res', 'libs', 'pdfjs', 'pdf.min.mjs'] },
             { 'path': ['res', 'html', 'static.js'] },
             { 'path': ['res', 'html', 'problem.css'] },
             { 'path': ['res', 'html', 'problem.js'] },
@@ -85,6 +89,9 @@ export default class {
             else if (file.path[file.path.length - 1].endsWith('.js')) {
                 return `<script src="${this.getRealPath(file.path)}"${attributes}></script>`;
             }
+            else if (file.path[file.path.length - 1].endsWith('.mjs')) {
+                return `<script type="module" src="${this.getRealPath(file.path)}"${attributes}></script>`;
+            }
             else {
                 throw new Error('Unknown file type');
             }
@@ -100,9 +107,12 @@ export default class {
         }).start().then(async (problemDetail) => {
             if (problemDetail?.json !== undefined) {
                 const problemContent = JSON.parse(problemDetail.json.pdoc.content);
-                const markdownContent: { [key: string]: string } = {};
+                const markdownContent: { [key: string]: { fetchData: { [key: string]: string }, content: string } } = {};
                 for (const [key, value] of Object.entries(problemContent)) {
-                    markdownContent[key] = await utils.parseMarkdown(value as string, pid);
+                    markdownContent[key] = await utils.parseMarkdown(this._extensionPath, this._panel.webview, value as string, `/d/${settings.domain}/p/${pid}/file`);
+                    for (const [id, url] of Object.entries(markdownContent[key].fetchData)) {
+                        this._tempFiles.push(id);
+                    }
                 }
                 const message = {
                     command: 'problem',
@@ -113,7 +123,7 @@ export default class {
                         markdownContent,
                     },
                 };
-                this._panel.webview.postMessage(message);
+                this.tryPostMessage(message);
             }
         }).catch(async (e: Error) => {
             io.error(e.message);
@@ -123,20 +133,42 @@ export default class {
             new fetch({ path: `/d/${settings.domain}/p/${pid}/solution`, addCookie: true }).start().then(async (solutionDetail) => {
                 if (solutionDetail?.json !== undefined) {
                     for (const solution of solutionDetail.json.psdocs) {
-                        solution.content = await utils.parseMarkdown(solution.content as string, pid);
+                        solution.content = await utils.parseMarkdown(this._extensionPath, this._panel.webview, solution.content as string, `/d/${settings.domain}/p/${pid}/file`);
+                        for (const [id, url] of Object.entries(solution.content.fetchData)) {
+                            this._tempFiles.push(id);
+                        }
                         for (const reply of solution.reply) {
-                            reply.content = await utils.parseMarkdown(reply.content as string, pid);
+                            reply.content = await utils.parseMarkdown(this._extensionPath, this._panel.webview, reply.content as string, `/d/${settings.domain}/p/${pid}/file`);
+                            for (const [id, url] of Object.entries(reply.content.fetchData)) {
+                                this._tempFiles.push(id);
+                            }
                         }
                     }
                     const message = {
                         command: 'solution',
                         data: solutionDetail.json,
                     };
-                    this._panel.webview.postMessage(message);
+                    this.tryPostMessage(message);
                 }
             }).catch(async (e: Error) => {
                 io.error(e.message);
             });
+        }
+    };
+
+    private tryPostMessage = (message: any) => {
+        try {
+            this._panel.webview.postMessage(message);
+        } catch (e) {
+            this.cleanup();
+        }
+    };
+
+    private cleanup = () => {
+        for (const id of this._tempFiles) {
+            const filePath = vscode.Uri.file(path.join(this._extensionPath, 'temp', id));
+            vscode.workspace.fs.delete(filePath);
+            outputChannel.info("Delete temp file", `"${filePath.toString()}"`);
         }
     };
 }